+++ src/main/java/nextstep/auth/authentication/token/TokenAuthenticationInterceptor.java
@@ -1,6 +1,9 @@
-package nextstep.auth.authentication;
+package nextstep.auth.authentication.token;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import nextstep.auth.authentication.AuthenticationException;
+import nextstep.auth.authentication.AuthenticationToken;
+import nextstep.auth.authentication.UserDetailService;
 import nextstep.auth.context.Authentication;
 import nextstep.auth.token.JwtTokenProvider;
 import nextstep.auth.token.TokenRequest;
